SSCS-10983 moves pdf functions into their own PdfRequestUtils service.
